requirements.txt
transformers==4.39.3
torch==2.2.2
torchvision==0.17.2
Pillow==10.2.0
opencv-python==4.9.0.80
tqdm==4.66.2
tk
huggingface_hub>=0.19.3,<1.0
diffusers==0.25.0
